refactor: split CAD initialization types
