+++ src/CADShared/Cad/Runtime/Initialization/AutoReflection.cs
@@ -3,90 +3,6 @@ namespace Fs.Fox.Cad;
 using System.Diagnostics;
 
 /// <summary>
-/// 加载时优先级
-/// </summary>
-[Flags]
-public enum Sequence : byte
-{
-    /// <summary>
-    /// 最先
-    /// </summary>
-    First,
-
-    /// <summary>
-    /// 最后
-    /// </summary>
-    Last,
-}
-
-/// <summary>
-/// 加载时自动执行接口
-/// </summary>
-public interface IFoxAutoGo
-{
-    /// <summary>
-    /// 控制加载顺序
-    /// </summary>
-    /// <returns></returns>
-    Sequence SequenceId();
-
-    /// <summary>
-    /// 关闭cad的时候会自动执行
-    /// </summary>
-    void Terminate();
-
-    /// <summary>
-    /// 打开cad的时候会自动执行
-    /// </summary>
-    void Initialize();
-}
-
-/// <summary>
-/// 加载时自动执行特性
-/// </summary>
-/// <remarks>
-/// 用于初始化和结束回收
-/// </remarks>
-/// <param name="sequence">优先级</param>
-/// <param name="isInitialize"><see langword="true"/>用于初始化;<see langword="false"/>用于结束回收</param>
-[AttributeUsage(AttributeTargets.Method)]
-// ReSharper disable once InconsistentNaming
-// ReSharper disable once ClassNeverInstantiated.Global
-public class IFoxInitializeAttribute(Sequence sequence = Sequence.Last, bool isInitialize = true) : Attribute
-{
-    /// <summary>
-    /// 优先级
-    /// </summary>
-    internal readonly Sequence SequenceId = sequence;
-
-    /// <summary>
-    /// <see langword="true"/>用于初始化;<see langword="false"/>用于结束回收
-    /// </summary>
-    internal readonly bool IsInitialize = isInitialize;
-}
-
-// 为了解决IExtensionApplication在一个dll内无法多次实现接口的关系
-// 所以在这里反射加载所有的 IAutoGo ,以达到能分开写"启动运行"函数的目的
-/// <summary>
-/// 执行此方法
-/// </summary>
-/// <param name="method"></param>
-/// <param name="sequence"></param>
-/// <param name="instance">已经创建的对象</param>
-internal class RunClass(MethodInfo method, Sequence sequence, object? instance = null)
-{
-    public Sequence Sequence { get; } = sequence;
-
-    /// <summary>
-    /// 运行方法
-    /// </summary>
-    public void Run()
-    {
-        method.Invoke(ref instance);
-    }
-}
-
-/// <summary>
 /// 此类作为加载后cad自动运行接口的一部分,用于反射特性和接口
 /// <para>
 /// 启动cad后的执行顺序为:<br/>
